implement book view

+++ src/main/java/com/mprog/springapp/dao/BookDao.java
@@ -4,5 +4,5 @@ import com.mprog.springapp.model.Book;
 import com.mprog.springapp.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookDao extends JpaRepository<Book, Long> {
+public interface BookDao extends JpaRepository<Book, Integer> {
 }
